Update README.md
The objective of the "Churned Client" project is to thoroughly evaluate a provided dataset. The primary focus is on analyzing the churn rate, understanding the reasons for churn, and examining the qualifications of clients who have churned. This analysis aims to provide deeper insights into the characteristics of churned clients.  The project includes an overview sheet that serves as a dashboard, presenting key findings in the form of charts and tables for easy comprehension. This dashboard offers a visual representation of the project's main insights.